Removed jwt directory as it has duplicate code in src.
The src directory was made of original files plus some files pulled from
other gir repos. The workflow was supposed to be to edit the original
files in jwt/ directory, call jwt/compile.sh, then copy the output to
src. The new workflow is to edit files in src except for the src/crypto
folder because files in that folder are pulled from external repos.